test(SingleSelect): move menu-popover regression test into SingleSelect.stories.tsx
Folds the InsideMenuPopover story into the main stories file and deletes the
separate spec.stories.tsx, matching the file's existing `step()` structure and
its use of `within(canvasElement.parentElement!)` to reach portalled content.

Also fixes the query that failed in CI: `SelectToggle` renders
`role="combobox"`, not a button, so `getByRole('button', { name: /Coffee/ })`
never resolved.

Chromatic is disabled for this story — it is interaction coverage, and the two
open popovers make for an unstable visual baseline.

Verified against a running Storybook: play function completes, the menu popover
survives selection, and the story reports no accessibility violations.

KZN-4210
